fix: get api 및 rename interface name
