+++ src/network/query/memory.ts
@@ -1,10 +1,7 @@
 import { useQuery } from 'react-query';
-import axios from 'axios';
+import { getForEntity } from '../Requests';
+import { Memory } from 'types/Memory';
 const memoryKey = 'memory';
 
-const getMemory = async () =>
-  await axios.get('http://localhost:3000/_mock/memory.json').then((res) => {
-    return res.data;
-  });
-
-export const useGetMemory = () => useQuery({ queryKey: [memoryKey], queryFn: () => getMemory() });
+export const useGetMemory = () =>
+  useQuery({ queryKey: [memoryKey], queryFn: (): Promise<Memory[]> => getForEntity('/api/memory', {}) });
