+++ src/types/Memory.ts
